# ui/openapi.json
{
  "components": {
    "schemas": {
      "CapabilitiesResponse": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "verdict_tags": {
            "items": {
              "type": "string"
            },
            "title": "Verdict Tags",
            "type": "array"
          }
        },
        "required": [
          "source",
          "verdict_tags"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
      "ClipItem": {
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "clip": {
            "title": "Clip",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "has_render_json": {
            "title": "Has Render Json",
            "type": "boolean"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "variant": {
            "title": "Variant",
            "type": "string"
          }
        },
        "required": [
          "clip",
          "variant",
          "size_bytes",
          "has_render_json",
          "fingerprint",
          "duration_s",
          "title",
          "caption"
        ],
        "title": "ClipItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "root": {
            "title": "Root",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "root",
          "schema_version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "SourceItem": {
        "properties": {
          "clip_count": {
            "title": "Clip Count",
            "type": "integer"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "poster_clip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Clip"
          },
          "poster_variant": {
            "title": "Poster Variant",
            "type": "string"
          },
          "stem": {
            "title": "Stem",
            "type": "string"
          },
          "variant_names": {
            "items": {
              "type": "string"
            },
            "title": "Variant Names",
            "type": "array"
          }
        },
        "required": [
          "stem",
          "display_name",
          "clip_count",
          "variant_names",
          "poster_clip",
          "poster_variant"
        ],
        "title": "SourceItem",
        "type": "object"
      },
      "SourceMetaIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "SourceMetaIn",
        "type": "object"
      },
      "SourceMetaOut": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "stem": {
            "title": "Stem",
            "type": "string"
          }
        },
        "required": [
          "stem",
          "display_name"
        ],
        "title": "SourceMetaOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "\u0410\u0432\u0442\u043e-\u0420\u0438\u043b\u0441 UI",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/capabilities": {
      "get": {
        "operationId": "capabilities_api_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Capabilities"
      }
    },
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/media/{stem}/{clip}": {
      "get": {
        "operationId": "media_api_media__stem___clip__get",
        "parameters": [
          {
            "in": "path",
            "name": "stem",
            "required": true,
            "schema": {
              "title": "Stem",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "clip",
            "required": true,
            "schema": {
              "title": "Clip",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "",
              "title": "Variant",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Media"
      }
    },
    "/api/sources": {
      "get": {
        "operationId": "sources_api_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceItem"
                  },
                  "title": "Response Sources Api Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sources"
      }
    },
    "/api/sources/{stem}/clips": {
      "get": {
        "operationId": "clips_api_sources__stem__clips_get",
        "parameters": [
          {
            "in": "path",
            "name": "stem",
            "required": true,
            "schema": {
              "title": "Stem",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipItem"
                  },
                  "title": "Response Clips Api Sources  Stem  Clips Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clips"
      }
    },
    "/api/sources/{stem}/meta": {
      "put": {
        "operationId": "set_source_meta_api_sources__stem__meta_put",
        "parameters": [
          {
            "in": "path",
            "name": "stem",
            "required": true,
            "schema": {
              "title": "Stem",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceMetaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMetaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Source Meta"
      }
    },
    "/api/thumb/{stem}/{clip}": {
      "get": {
        "operationId": "thumb_api_thumb__stem___clip__get",
        "parameters": [
          {
            "in": "path",
            "name": "stem",
            "required": true,
            "schema": {
              "title": "Stem",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "clip",
            "required": true,
            "schema": {
              "title": "Clip",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "",
              "title": "Variant",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thumb"
      }
    }
  }
}